feat: update object-to-ast to TS (work-in-progress)
